requirements.txt
flask
google-genai
requests
gunicorn
mangum
python-dotenv
boto3